requirements.txt
Django==2.1
django-filter==2.0.0
djangorestframework==3.8.2
Markdown==2.6.11
pytz==2018.5
requests==2.19.1
coreapi==2.3.3
PyYAML==3.12
psycopg2-binary==2.7.5
gunicorn